Delete complété

+++ src/app/liste-livres/liste-livres.ts
@@ -1,4 +1,3 @@
-// liste-livres.ts
 import { Component, OnInit, inject, ChangeDetectorRef } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { ActivatedRoute } from '@angular/router';
@@ -6,7 +5,6 @@ import { Observable } from 'rxjs';
 import { Book } from '../book';
 import { BookService } from '../book-service';
 import { Router } from '@angular/router';
-import { error } from 'console';
 
 @Component({
   selector: 'app-liste-livres',
@@ -21,34 +19,45 @@ export class ListeLivres implements OnInit {
   private router = inject(Router);
   private cdr = inject(ChangeDetectorRef);
   
-  // Observable pour les livres
   books$!: Observable<Book[]>;
   loading = false;
 
   ngOnInit(): void {
     this.loading = true;
     this.route.params.subscribe(() => {
-      this.books$ = this.bookService.getListeLivre();
+      this.loadBooks();
     });
   }
 
+  loadBooks()
+  {
+    this.loading = true;
+    this.books$ = this.bookService.getListeLivre();
+    this.loading = false;
+  }
 
   updateBook(id: number)
   {
-      this.router.navigate(['update-book', id]);
+    this.router.navigate(['update-book', id]);
   }
 
   deleteBook(id: number)
   {
+    if(!confirm('Voulez-vous vraiment supprimer ce livre ?')) //Confirm() permet de poser une question avec deux boutons.
+    {
+      return; //Si la confirmation est annulé, on ne fait rien.
+    }
+
     this.bookService.deleteLivre(id).subscribe({
-      next: (data) =>
-      {
+      next: (data) => {
+        console.log('Livre supprimé :', data);
+        this.loadBooks();
         this.cdr.detectChanges();
-        console.log(data);
+        
       },
-      error: (error) =>
-      {
-        console.error('Erreur lors de la suppression du livre :', error);
+      error: (error) => {
+        console.error('Erreur :', error);
+        alert('Erreur lors de la suppression du livre.');
       }
     });
   }
